seperating navbar into seperate files
