+++ client/pages/BoardContext.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const BoardContext = React.createContext();
+
+export default BoardContext;
